# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FixturePrediction": {
        "properties": {
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "match": {
            "$ref": "#/components/schemas/MatchSummary"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerFixturePrediction"
            },
            "title": "Players",
            "type": "array"
          },
          "prediction_cutoff": {
            "format": "date-time",
            "title": "Prediction Cutoff",
            "type": "string"
          },
          "prediction_run_id": {
            "format": "uuid",
            "title": "Prediction Run Id",
            "type": "string"
          }
        },
        "required": [
          "match",
          "prediction_run_id",
          "prediction_cutoff",
          "data_cutoff",
          "players"
        ],
        "title": "FixturePrediction",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "PlayerDetail": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "current_team": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nationality_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality Code"
          },
          "photo_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Source"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "preferred_foot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Foot"
          },
          "primary_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Position"
          }
        },
        "required": [
          "id",
          "display_name",
          "primary_position",
          "nationality_code",
          "active",
          "date_of_birth",
          "preferred_foot",
          "height_cm"
        ],
        "title": "PlayerDetail",
        "type": "object"
      },
      "PlayerFixtureOutlook": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "expected_points": {
            "title": "Expected Points",
            "type": "number"
          },
          "explanations": {
            "items": {
              "$ref": "#/components/schemas/ExplanationView"
            },
            "title": "Explanations",
            "type": "array"
          },
          "football": {
            "$ref": "#/components/schemas/PlayerFootballOutlook"
          },
          "match": {
            "$ref": "#/components/schemas/MatchSummary"
          },
          "median_points": {
            "title": "Median Points",
            "type": "number"
          },
          "p10_points": {
            "title": "P10 Points",
            "type": "number"
          },
          "p90_points": {
            "title": "P90 Points",
            "type": "number"
          },
          "scoring_appearance_probability": {
            "description": "Probability of positive minutes meeting the league's appearance_minimum_minutes.",
            "title": "Scoring Appearance Probability",
            "type": "number"
          }
        },
        "required": [
          "match",
          "expected_points",
          "median_points",
          "p10_points",
          "p90_points",
          "available",
          "scoring_appearance_probability",
          "confidence",
          "football",
          "explanations"
        ],
        "title": "PlayerFixtureOutlook",
        "type": "object"
      },
      "PlayerFixturePrediction": {
        "properties": {
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "match": {
            "$ref": "#/components/schemas/MatchSummary"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "prediction_cutoff": {
            "format": "date-time",
            "title": "Prediction Cutoff",
            "type": "string"
          },
          "prediction_run_id": {
            "format": "uuid",
            "title": "Prediction Run Id",
            "type": "string"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/PredictionValue"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "prediction_run_id",
          "match",
          "prediction_cutoff",
          "data_cutoff",
          "values"
        ],
        "title": "PlayerFixturePrediction",
        "type": "object"
      },
...
      "PlayerOutlookRequest": {
        "properties": {
          "horizon": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          }
        },
        "title": "PlayerOutlookRequest",
        "type": "object"
      },
      "PlayerOutlookView": {
        "properties": {
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "decision_cutoff": {
            "format": "date-time",
            "title": "Decision Cutoff",
            "type": "string"
          },
          "fixtures": {
            "items": {
              "$ref": "#/components/schemas/PlayerFixtureOutlook"
            },
            "title": "Fixtures",
            "type": "array"
          },
          "league_id": {
            "format": "uuid",
            "title": "League Id",
            "type": "string"
          },
          "model_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Model Versions",
            "type": "object"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "prediction_cutoff": {
            "format": "date-time",
            "title": "Prediction Cutoff",
            "type": "string"
          },
          "prediction_run_id": {
            "format": "uuid",
            "title": "Prediction Run Id",
            "type": "string"
          },
          "recommendation_score": {
            "$ref": "#/components/schemas/PlayerRecommendationScore"
          },
          "requested_horizon": {
            "title": "Requested Horizon",
            "type": "integer"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "rules_version": {
            "title": "Rules Version",
            "type": "integer"
          },
          "scoring_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Role"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "simulation_count": {
            "title": "Simulation Count",
            "type": "integer"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "league_id",
          "rules_version",
          "roles",
          "scoring_role",
          "recommendation_score",
          "requested_horizon",
          "fixtures",
          "prediction_run_id",
          "model_versions",
          "prediction_cutoff",
          "data_cutoff",
          "decision_cutoff",
          "simulation_count",
          "seed"
        ],
        "title": "PlayerOutlookView",
        "type": "object"
      },
      "PlayerPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "PlayerPage",
        "type": "object"
      },
      "PlayerRecentFormView": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "coverage": {
            "const": "observed_player_matches",
            "default": "observed_player_matches",
            "title": "Coverage",
            "type": "string"
          },
          "data_cutoff": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Cutoff"
          },
          "items": {
            "description": "Newest observed finished player matches first; missing matches are not zeroes.",
            "items": {
              "$ref": "#/components/schemas/PlayerFormMatch"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "as_of",
          "data_cutoff",
          "limit",
          "items"
        ],
        "title": "PlayerRecentFormView",
        "type": "object"
      },
      "PlayerRecommendationScore": {
        "properties": {
          "objective": {
            "const": "expected_points",
            "default": "expected_points",
            "title": "Objective",
            "type": "string"
          },
          "scope": {
            "const": "individual_next_fixture",
            "default": "individual_next_fixture",
            "title": "Scope",
            "type": "string"
          },
          "unit": {
            "const": "fantasy_points",
            "default": "fantasy_points",
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "description": "Mean individual fantasy points in the next fixture under the active league rules. Not a 0-100 rating, a probability, or a guarantee of selection in a legal XI. Lineup-level modifiers and substitutions are excluded.",
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value"
        ],
        "title": "PlayerRecommendationScore",
        "type": "object"
      },
...
      "PlayerSummary": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nationality_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality Code"
          },
          "photo_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Source"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "primary_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Position"
          }
        },
        "required": [
          "id",
          "display_name",
          "primary_position",
          "nationality_code",
          "active"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
      "PredictionValue": {
        "properties": {
          "expected_value": {
            "title": "Expected Value",
            "type": "number"
          },
          "median": {
            "title": "Median",
            "type": "number"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "p10": {
            "title": "P10",
            "type": "number"
          },
          "p90": {
            "title": "P90",
            "type": "number"
          },
          "probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability"
          },
          "reliability": {
            "title": "Reliability",
            "type": "number"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "model_version",
          "expected_value",
          "median",
          "p10",
          "p90",
          "probability",
          "reliability"
        ],
        "title": "PredictionValue",
        "type": "object"
      },
...
      "TeamSummary": {
        "properties": {
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          }
        },
        "required": [
          "id",
          "name",
          "short_name",
          "country_code"
        ],
        "title": "TeamSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Probabilistic decision support for Italian Fantacalcio.",
    "title": "MarkGuardiola",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/leagues/{league_id}/players/{player_id}/outlook": {
      "post": {
        "operationId": "getPlayerLeagueOutlook",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "League Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerOutlookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerOutlookView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Player League Outlook",
        "tags": [
          "player-insights"
        ]
      }
    },
...
    "/api/v1/players": {
      "get": {
        "operationId": "listPlayers",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "position",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Players",
        "tags": [
          "football"
        ]
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "operationId": "getPlayer",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player",
        "tags": [
          "football"
        ]
      }
    },
    "/api/v1/players/{player_id}/recent-form": {
      "get": {
        "operationId": "getPlayerRecentForm",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRecentFormView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Player Recent Form",
        "tags": [
          "player-insights"
        ]
      }
    },
    "/api/v1/predictions/fixture/{match_id}": {
      "get": {
        "operationId": "getFixturePredictions",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixturePrediction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fixture Predictions",
        "tags": [
          "predictions"
        ]
      }
    },
    "/api/v1/predictions/player/{player_id}": {
      "get": {
        "operationId": "getPlayerPredictions",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayerFixturePrediction"
                  },
                  "title": "Response Getplayerpredictions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Player Predictions",
        "tags": [
          "predictions"
        ]
      }
    },
...
  }
}